In response to PR comments, merged study_submit.py python scripts into one, trialling sanitizeOutput for test snapshot.
