+++ .gitignore
@@ -10,5 +10,3 @@ null/
 .nf-test*
 .idea/
 test_data
-.claude/
-CLAUDE.md
